requirements.txt
flask
python-dotenv
requests
SmartApi
